requirements.txt
Django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0